--- src/js/chessground/index.ts
@@ -1,13 +1,0 @@
-import Chessground from './Chessground'
-import fen from './fen'
-import * as util from './util'
-import * as board from './board'
-import drag from './drag'
-
-export default {
-  Chessground,
-  fen,
-  util,
-  board,
-  drag
-}
